Added The structure files and first release.
This Is the idea I envisioned. Future changes may include other search options, ability to change the default search from Brave, and a side bar for shortcuts, as there is always a margin on the left regardless of how small the window.  I'm considering adding the microsoft office dropbox if it's possible. Maybe even a post-it widget.